Update the fusd examples now that multiple registration is no longer bitrotted.
git-svn-id: http://svn.xiph.org/trunk/fusd@12355 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ examples/drums.c
@@ -59,7 +59,7 @@
 static char *drums_strings[] = {"bam", "bum", "beat", "boom",
 				"bang", "crash", NULL};
 
-int drums_read(struct fusd_file_info *file, char *user_buffer,
+ssize_t drums_read(struct fusd_file_info *file, char *user_buffer,
 	       size_t user_length, loff_t *offset)
 {
   int len;
